refactor(mastracode): decouple non-TUI code from src/tui
Move TUI-independent modules out of src/tui so the upcoming package split
(core/tui/web) has a clean boundary:

- mastra brand palette -> src/theme-palette.ts (index.ts mode colors no
  longer import tui/theme; palette values are read before theme adaptation
  runs, so behavior is unchanged for green/orange; purple mode color now
  uses the raw brand constant)
- tui/voice/stt-registry -> src/voice/stt-registry (used by onboarding
  settings)
- settings-mutation helpers out of tui commands into onboarding/
  (custom-providers, custom-packs, om-settings); tui commands re-export
- onboarding-inline (a TUI component) -> src/tui/onboarding-inline
