+++ mastracode/src/onboarding/index.ts
@@ -1,5 +1,3 @@
-export { OnboardingInlineComponent } from './onboarding-inline.js';
-export type { OnboardingResult, OnboardingOptions } from './onboarding-inline.js';
 export { getAvailableModePacks, getAvailableOmPacks, ONBOARDING_VERSION } from './packs.js';
 export type { ModePack, OMPack, ProviderAccess, ProviderAccessLevel } from './packs.js';
 export {
