+++ mastracode/src/tui/commands/models-pack.ts
@@ -3,6 +3,11 @@ import type { SelectItem } from '@earendil-works/pi-tui';
 import chalk from 'chalk';
 
 import { setClipboardText } from '../../clipboard/index.js';
+import { removeCustomPackFromSettings } from '../../onboarding/custom-packs.js';
+
+// Re-exported for existing importers/tests; the implementation lives in
+// onboarding/custom-packs.ts so non-TUI surfaces (web routes) can share it.
+export { removeCustomPackFromSettings };
 import type { ModePack, ProviderAccess, ProviderAccessLevel } from '../../onboarding/packs.js';
 import { getAvailableModePacks } from '../../onboarding/packs.js';
 import {
@@ -349,30 +354,6 @@ export function upsertCustomPackInSettings(
   if (setActive) {
     settings.models.activeModelPackId = pack.id;
     settings.models.modeDefaults = modeDefaults;
-  }
-}
-
-export function removeCustomPackFromSettings(settings: GlobalSettings, packId: string): void {
-  if (!packId.startsWith('custom:')) return;
-  const packName = packId.slice('custom:'.length);
-  const removedPack = settings.customModelPacks.find(p => p.name === packName);
-  settings.customModelPacks = settings.customModelPacks.filter(p => p.name !== packName);
-
-  const modeDefaultsMatchRemovedPack =
-    !!removedPack &&
-    settings.models.modeDefaults.plan === removedPack.models.plan &&
-    settings.models.modeDefaults.build === removedPack.models.build &&
-    settings.models.modeDefaults.fast === removedPack.models.fast;
-
-  if (settings.models.activeModelPackId === packId) {
-    settings.models.activeModelPackId = null;
-    settings.models.modeDefaults = {};
-  } else if (modeDefaultsMatchRemovedPack) {
-    settings.models.modeDefaults = {};
-  }
-
-  if (settings.onboarding.modePackId === packId) {
-    settings.onboarding.modePackId = null;
   }
 }
