+++ mastracode/src/voice/__tests__/stt-registry.test.ts
